# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PRIN135-Demo-10-API-Documentation",
    "version": "1.0",
    "description": "API Documentation for Task Activity",
    "contact": {
      "email": "myemail@sample.com",
      "name": "First Last"
    }
  },
  "externalDocs": {
    "url": "openapi.json",
    "description": "External documentation link."
  },
  "servers": [
    {
      "url": "https://demo-09-db-integration.vercel.app",
      "variables": {},
      "description": "Production"
    }
  ],
  "paths": {
    "/tasks": {
      "summary": "Tasks base path.",
      "description": "Tasks base path.",
      "get": {
        "summary": "Returns all tasks.",
        "description": "Returns all tasks.",
        "operationId": "GetAllTasks",
        "responses": {
          "200": {
            "description": "Tasks found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskGetResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "post": {
        "summary": "Creates a task.",
        "description": "Creates a task.",
        "operationId": "CreateTask",
        "requestBody": {
          "description": "To create a task, add values to the request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created."
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{taskID}": {
      "summary": "Tasks base path with taskID.",
      "description": "Tasks base path with taskID.",
      "get": {
        "summary": "Returns a task by TaskID.",
        "description": "Returns a task by TaskID.",
        "operationId": "GetTask",
        "responses": {
          "200": {
            "description": "Task found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskGetResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "put": {
        "summary": "Updates a task by TaskID.",
        "description": "Updates a task by TaskID.",
        "operationId": "UpdateTask",
        "requestBody": {
          "description": "To create a task, add values to the request body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "summary": "Deletes a task by TaskID.",
        "description": "Deletes a task by TaskID.",
        "operationId": "DeleteTask",
        "responses": {
          "200": {
            "description": "Task deleted."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "parameters": [
        {
          "name": "taskID",
          "in": "path",
          "description": "Task ID",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "TaskCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isdone": {
            "type": "boolean"
          }
        }
      },
      "TaskGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "isdone": {
            "type": "boolean"
          }
        }
      },
      "TaskUpdateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "isdone": {
            "type": "boolean"
          }
        }
      }
    }
  }
}